break out `replaceOrCreateFrontmatter`
