--- lib/rule-doc-title-format.ts
@@ -1,9 +1,0 @@
-export const RULE_DOC_TITLE_FORMATS = [
-  'desc',
-  'desc-parens-name',
-  'desc-parens-prefix-name',
-  'name',
-  'prefix-name',
-] as const;
-
-export type RuleDocTitleFormat = (typeof RULE_DOC_TITLE_FORMATS)[number];
