+++ test/fixtures/cjs-config-extends/README.md
@@ -9,9 +9,9 @@
 
 | Name                           | Description            | 💼 |
 | :----------------------------- | :--------------------- | :- |
-| [no-bar](docs/rules/no-bar.md) | Description of no-bar. | ✅ |
-| [no-baz](docs/rules/no-baz.md) | Description of no-baz. | ✅ |
-| [no-biz](docs/rules/no-biz.md) | Description of no-biz. | ✅ |
-| [no-foo](docs/rules/no-foo.md) | Description of no-foo. | ✅ |
+| [no-bar](docs/rules/no-bar.md) | Description of no-bar. | ✅  |
+| [no-baz](docs/rules/no-baz.md) | Description of no-baz. | ✅  |
+| [no-biz](docs/rules/no-biz.md) | Description of no-biz. | ✅  |
+| [no-foo](docs/rules/no-foo.md) | Description of no-foo. | ✅  |
 
 <!-- end auto-generated rules list -->
